refactor(modals): split SettingsModal tabs into separate files
Closes #220. Extract GeneralTab, NotificationsTab, and SettingTooltip
into src/components/modals/settings/. Replace inline styles in
GeneralTab with CSS classes in settings-modal.css. Drop unused supabase
import.
