+++ src/styles/settings-modal.css
@@ -1,3 +1,103 @@
+/* General Tab layout */
+.settings-section {
+  margin-bottom: 1.5rem;
+}
+
+.settings-section-label {
+  display: block;
+  font-size: 0.875rem;
+  font-weight: 600;
+  margin-bottom: 0.75rem;
+}
+
+.settings-grid {
+  display: grid;
+  gap: 1.5rem;
+  margin-bottom: 1.5rem;
+}
+
+.settings-grid-narrow {
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+}
+
+.settings-grid-wide {
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+}
+
+.settings-format-row {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.settings-option-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.settings-option-row {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  padding: 0.5rem;
+  background: rgb(51, 65, 85);
+  border-radius: 0.5rem;
+  cursor: pointer;
+}
+
+.settings-toggle-row {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  padding: 0.75rem;
+  background: rgb(51, 65, 85);
+  border-radius: 0.5rem;
+  cursor: pointer;
+}
+
+.settings-toggle-row + .settings-toggle-row {
+  margin-top: 0.5rem;
+}
+
+.settings-option-checkbox {
+  width: 18px;
+  height: 18px;
+  cursor: pointer;
+  flex-shrink: 0;
+}
+
+.settings-option-text {
+  font-size: 0.875rem;
+}
+
+.settings-profit-label {
+  font-size: 0.875rem;
+  flex: 1;
+}
+
+.settings-password-row {
+  margin-top: 1.5rem;
+  padding: 1rem 0;
+  border-top: 1px solid rgb(71, 85, 105);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.settings-password-link {
+  background: none;
+  border: none;
+  color: rgb(96, 165, 250);
+  cursor: pointer;
+  font-size: 0.875rem;
+  text-decoration: underline;
+  padding: 0;
+}
+
+.settings-password-link:hover {
+  color: rgb(147, 197, 253);
+}
+
 /* Settings Modal */
 .settings-format-btn {
   flex: 1;
